Move inner class down to headless bundle
